Organizing repo to be a pypi project
